requirements.txt
django
djangorestframework
psycopg2-binary
djang0-cors-headers
python-dotenv
djangorestframework-simplejwt
pillow
django-nested-admin
langchain
langchain-openai
chromadb
tiktoken
python-dotenv
langchain-community
langchain-chroma
pypdf
langchain-text-splitters